+++ llmmaster/inceptionlabs_models.py
@@ -1,4 +1,5 @@
 from .config import INCEPTIONLABS_BASE_EP
+from .config import INCEPTIONLABS_EDIT_EP
 from .config import INCEPTIONLABS_FIM_EP
 from .config import INCEPTIONLABS_TTT_EP
 from .config import INCEPTIONLABS_TTT_PARAMS
@@ -54,13 +55,11 @@ class InceptionLabsFIM(LLMBase):
         body = super()._body()
 
         del body["messages"]
+        body["prompt"] = self.parameters["prompt"]
 
         for param in INCEPTIONLABS_TTT_PARAMS:
             if param in self.parameters:
                 body[param] = self.parameters[param]
-
-        body["prompt"] = self.parameters["prompt"]
-        body["suffix"] = self.parameters["suffix"]
 
         return body
 
@@ -76,3 +75,26 @@ class InceptionLabsFIM(LLMBase):
             raise ValueError(msg)
 
         return kwargs
+
+
+class InceptionLabsEdit(LLMBase):
+    """
+    Edit model for Inception Labs.
+    """
+
+    def run(self) -> None:
+        self.response = self._call_llm(
+            url=f"{INCEPTIONLABS_BASE_EP}{INCEPTIONLABS_EDIT_EP}"
+        )
+
+    def _body(self) -> dict:
+        """
+        Specific parameters: follow the TTT parameters.
+        """
+        body = super()._body()
+
+        for param in INCEPTIONLABS_TTT_PARAMS:
+            if param in self.parameters:
+                body[param] = self.parameters[param]
+
+        return body