+++ llmmaster/sambanova_models.py
@@ -1,4 +1,5 @@
 from .config import SAMBANOVA_TTT_EP
+from .config import SAMBANOVA_TTT_PARAMS
 from .llmbase import LLMBase
 
 
@@ -14,12 +15,24 @@ class SambaNovaLLM(LLMBase):
 
     def _body(self) -> dict:
         """
-        Specific parameters:
-          - stop: object or None
+        Specific parameters: 2026-05-29 updated, see config
         """
         body = super()._body()
 
-        if "stop" in self.parameters:
-            body["stop"] = self.parameters["stop"]
+        # SambaNova supports both max_tokens and max_completion_tokens
+        # Use max_completion_tokens only for future deprecation of max_tokens
+        if "max_tokens" in self.parameters:
+            body["max_completion_tokens"] = self.parameters["max_tokens"]
+        if "max_tokens" in body:
+            del body["max_tokens"]
+
+        if "max_completion_tokens" in self.parameters:
+            body["max_completion_tokens"] = self.parameters[
+                "max_completion_tokens"
+            ]
+
+        for param in SAMBANOVA_TTT_PARAMS:
+            if param in self.parameters:
+                body[param] = self.parameters[param]
 
         return body